// scripts/injective/proto/namespace_pb.ts
export interface Namespace {
    /**
     * The tokenfactory denom to which this namespace applies to
     *
     * @generated from protobuf field: string denom = 1
     */
    denom: string;
    /**
     * The address of cosmwasm contract to apply code-based restrictions
     *
     * @generated from protobuf field: string wasm_hook = 2
     */
    wasmHook: string;
    /**
     * permissions for each role
     *
     * @generated from protobuf field: repeated injective.permissions.v1beta1.Role role_permissions = 3
     */
    rolePermissions: Role[];
    /**
     * roles for each actor
     *
     * @generated from protobuf field: repeated injective.permissions.v1beta1.ActorRoles actor_roles = 4
     */
    actorRoles: ActorRoles[];
    /**
     * managers for each role
     *
     * @generated from protobuf field: repeated injective.permissions.v1beta1.RoleManager role_managers = 5
     */
    roleManagers: RoleManager[];
    /**
     * status for each policy
     *
     * @generated from protobuf field: repeated injective.permissions.v1beta1.PolicyStatus policy_statuses = 6
     */
    policyStatuses: PolicyStatus[];
    /**
     * capabilities for each manager for each policy
     *
     * @generated from protobuf field: repeated injective.permissions.v1beta1.PolicyManagerCapability policy_manager_capabilities = 7
     */
    policyManagerCapabilities: PolicyManagerCapability[];
    /**
     * The address of the EVM contract to map code-based permissions
     *
     * @generated from protobuf field: string evm_hook = 8
     */
    evmHook: string;
}
/**
 * Role defines permissions for a specific role
 *
 * @generated from protobuf message injective.permissions.v1beta1.Role
 */
export interface Role {
    /**
     * Role name
     *
     * @generated from protobuf field: string name = 1
     */
    name: string;
    /**
     * Role ID
     *
     * @generated from protobuf field: uint32 role_id = 2
     */
    roleId: number;
    /**
     * Permission bitmap
     *
     * @generated from protobuf field: uint32 permissions = 3
     */
    permissions: number;
}
/**
 * ActorRoles defines roles for an actor
 *
 * @generated from protobuf message injective.permissions.v1beta1.ActorRoles
 */
export interface ActorRoles {
    /**
     * The actor name
     *
     * @generated from protobuf field: string actor = 1
     */
    actor: string;
    /**
     * The roles for the actor
     *
     * @generated from protobuf field: repeated string roles = 2
     */
    roles: string[];
}
/**
 * RoleManager defines which roles a manager can manage
 *
 * @generated from protobuf message injective.permissions.v1beta1.RoleManager
 */
export interface RoleManager {
    /**
     * The manager address
     *
     * @generated from protobuf field: string manager = 1
     */
    manager: string;
    /**
     * The roles this manager can manage
     *
     * @generated from protobuf field: repeated string roles = 2
     */
    roles: string[];
}
/**
 * PolicyStatus defines the status of a policy action
 *
 * @generated from protobuf message injective.permissions.v1beta1.PolicyStatus
 */
export interface PolicyStatus {
    /**
     * The action permission bit
     *
     * @generated from protobuf field: uint32 action = 1
     */
    action: number;
    /**
     * Whether the action is disabled
     *
     * @generated from protobuf field: bool is_disabled = 2
     */
    isDisabled: boolean;
    /**
     * Whether the action is sealed (permanently locked)
     *
     * @generated from protobuf field: bool is_sealed = 3
     */
    isSealed: boolean;
}
/**
 * PolicyManagerCapability defines manager capabilities for policies
 *
 * @generated from protobuf message injective.permissions.v1beta1.PolicyManagerCapability
 */
export interface PolicyManagerCapability {
    /**
     * The manager address
     *
     * @generated from protobuf field: string manager = 1
     */
    manager: string;
    /**
     * The policy action
     *
     * @generated from protobuf field: uint32 action = 2
     */
    action: number;
    /**
     * Whether manager can enable/disable
     *
     * @generated from protobuf field: bool can_toggle = 3
     */
    canToggle: boolean;
    /**
     * Whether manager can seal
     *
     * @generated from protobuf field: bool can_seal = 4
     */
    canSeal: boolean;
}
/**
 * MsgCreateNamespace is the message to create a new namespace
 *
 * @generated from protobuf message injective.permissions.v1beta1.MsgCreateNamespace
 */
export interface MsgCreateNamespace {
    /**
     * The sender's Injective address
     *
     * @generated from protobuf field: string sender = 1
     */
    sender: string;
    /**
     * The namespace information
     *
     * @generated from protobuf field: injective.permissions.v1beta1.Namespace namespace = 2
     */
    namespace?: Namespace;
}
// ...
export const Namespace = new Namespace$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Role$Type extends MessageType<Role> {
    constructor() {
        super("injective.permissions.v1beta1.Role", [
            { no: 1, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "role_id", kind: "scalar", T: 13 /*ScalarType.UINT32*/ },
            { no: 3, name: "permissions", kind: "scalar", T: 13 /*ScalarType.UINT32*/ }
        ]);
    }
    create(value?: PartialMessage<Role>): Role {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.name = "";
        message.roleId = 0;
        message.permissions = 0;
        if (value !== undefined)
            reflectionMergePartial<Role>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Role): Role {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string name */ 1:
                    message.name = reader.string();
                    break;
                case /* uint32 role_id */ 2:
                    message.roleId = reader.uint32();
                    break;
                case /* uint32 permissions */ 3:
                    message.permissions = reader.uint32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Role, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string name = 1; */
        if (message.name !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.name);
        /* uint32 role_id = 2; */
        if (message.roleId !== 0)
            writer.tag(2, WireType.Varint).uint32(message.roleId);
        /* uint32 permissions = 3; */
        if (message.permissions !== 0)
            writer.tag(3, WireType.Varint).uint32(message.permissions);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message injective.permissions.v1beta1.Role
 */
export const Role = new Role$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ActorRoles$Type extends MessageType<ActorRoles> {
    constructor() {
        super("injective.permissions.v1beta1.ActorRoles", [
            { no: 1, name: "actor", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "roles", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<ActorRoles>): ActorRoles {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.actor = "";
        message.roles = [];
        if (value !== undefined)
            reflectionMergePartial<ActorRoles>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ActorRoles): ActorRoles {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string actor */ 1:
                    message.actor = reader.string();
                    break;
                case /* repeated string roles */ 2:
                    message.roles.push(reader.string());
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ActorRoles, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string actor = 1; */
        if (message.actor !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.actor);
        /* repeated string roles = 2; */
        for (let i = 0; i < message.roles.length; i++)
            writer.tag(2, WireType.LengthDelimited).string(message.roles[i]);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message injective.permissions.v1beta1.ActorRoles
 */
export const ActorRoles = new ActorRoles$Type();
// ...
export const RoleManager = new RoleManager$Type();
// ...
export const PolicyStatus = new PolicyStatus$Type();
// ...
export const PolicyManagerCapability = new PolicyManagerCapability$Type();
// @generated message type with reflection information, may provide speed optimized methods
class MsgCreateNamespace$Type extends MessageType<MsgCreateNamespace> {
    constructor() {
        super("injective.permissions.v1beta1.MsgCreateNamespace", [
            { no: 1, name: "sender", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "namespace", kind: "message", T: () => Namespace }
        ]);
    }
    create(value?: PartialMessage<MsgCreateNamespace>): MsgCreateNamespace {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.sender = "";
        if (value !== undefined)
            reflectionMergePartial<MsgCreateNamespace>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: MsgCreateNamespace): MsgCreateNamespace {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string sender */ 1:
                    message.sender = reader.string();
                    break;
                case /* injective.permissions.v1beta1.Namespace namespace */ 2:
                    message.namespace = Namespace.internalBinaryRead(reader, reader.uint32(), options, message.namespace);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: MsgCreateNamespace, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string sender = 1; */
        if (message.sender !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.sender);
        /* injective.permissions.v1beta1.Namespace namespace = 2; */
        if (message.namespace)
            Namespace.internalBinaryWrite(message.namespace, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message injective.permissions.v1beta1.MsgCreateNamespace
 */
export const MsgCreateNamespace = new MsgCreateNamespace$Type();